refactor: decompose AccessibilityServiceClient into delegate modules
Decompose the 4,652-line AccessibilityServiceClient.ts into 7 focused
delegate modules using composition pattern. The main client becomes an
orchestrator that holds delegates and implements the AccessibilityService
interface by delegating to modules.

New module structure:
- AccessibilityServiceGestures: swipe, tap, drag, pinch operations
- AccessibilityServiceText: setText, clearText, imeAction, selectAll
- AccessibilityServiceHierarchy: hierarchy retrieval, caching, conversion
- AccessibilityServiceStorage: SharedPreferences operations
- AccessibilityServiceCertificates: CA cert install/remove, permissions
- AccessibilityServiceFocus: TalkBack focus, traversal order
- AccessibilityServiceHighlights: visual highlight overlays
- types.ts: DelegateContext interfaces and shared types

Maintains backward compatibility via re-exports at the original location.

Closes #1107
